fixed last windows warning for SILACAnalyzer.C
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.8@8164 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/APPLICATIONS/TOPP/SILACAnalyzer.C
@@ -460,7 +460,7 @@ class TOPPSILACAnalyzer
     {
       for (UInt j = 0; j < SILAClabels[i].size(); ++j)
       {
-        Int found = labels.find(SILAClabels[i][j]);
+        Int found = (Int) labels.find(SILAClabels[i][j]);
 
         if (found < 0)
         {
